+++ Hazard/src/Hazard/Rendering/Batch.h
@@ -8,7 +8,7 @@ namespace Hazard::Rendering
 	class Batch {
 	public:
 		Batch() : m_DataSize(sizeof(Type)) {}
-		Batch(size_t size) : m_DataSize(sizeof(Type)) {
+		Batch(uint32_t size) : m_DataSize(sizeof(Type)) {
 			Resize(size);
 		};
 		~Batch() {
@@ -26,7 +26,7 @@ namespace Hazard::Rendering
 			m_IndexCount = 0;
 			m_Count = 0;
 		};
-		void Resize(size_t size) 
+		void Resize(uint32_t size) 
 		{
 			m_Size = size;
 			m_BufferSize = m_Size * m_DataSize;
@@ -38,8 +38,8 @@ namespace Hazard::Rendering
 		void* GetData() { return m_BufferBase; }
 		uint32_t GetCount() { return m_Count; }
 		uint32_t GetSize() { return m_Size; }
-		size_t GetDataSize() { return (uint32_t)((uint8_t*)m_BufferPointer - (uint8_t*)m_BufferBase); }
-		size_t GetIndexCount() { return m_IndexCount; }
+		uint32_t GetDataSize() { return (uint32_t)((uint8_t*)m_BufferPointer - (uint8_t*)m_BufferBase); }
+		uint32_t GetIndexCount() { return m_IndexCount; }
 		operator bool() const { return m_IndexCount != 0; }
 
 	private:
